fixed some bugs for the survey backend. The registration flow is broken currently as putUserDemographics is not sending data to the database.

+++ lib/Onboarding/Issues/Climate.dart
@@ -35,8 +35,8 @@ class _ClimateState extends State<Climate> {
   void initState() {
     super.initState();
     setState(() {
-      alignRating = widget.ratings.climateScore;
-      valueRating = widget.ratings.climateWeight;
+      alignRating = widget.ratings.climateScore.toDouble();
+      valueRating = widget.ratings.climateWeight.toDouble();
     });
     updateButton();
   }
